Octave AI v1/octave-ai/backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
requests==2.31.0
groq>=0.11.0
google-generativeai==0.3.2
openai==1.3.0
elevenlabs==0.2.26
azure-cognitiveservices-speech==1.34.0
pymongo==4.6.0